Rename the module name to ui-fixtures-plugin
