--- paywall-fixtures-plugin/gradle.properties
@@ -1,2 +1,0 @@
-POM_ARTIFACT_ID=purchases-paywall-fixtures-plugin
-POM_NAME=RevenueCat Paywall Fixtures Plugin
