+++ settings.gradle.kts
@@ -65,7 +65,7 @@ include(":ui:revenuecatui")
 include(":ui:revenuecatui-testing")
 include(":bom")
 include(":codegen")
-include(":paywall-fixtures-plugin")
+include(":ui-fixtures-plugin")
 include(":examples:paywall-tester")
 include(":test-apps:testpurchasesandroidcompatibility")
 include(":test-apps:testpurchasesuiandroidcompatibility")
