+++ ui-fixtures-plugin/gradle.properties
@@ -1,0 +1,2 @@
+POM_ARTIFACT_ID=purchases-ui-fixtures-plugin
+POM_NAME=RevenueCat UI Fixtures Plugin
