+++ ui-fixtures-plugin/src/test/kotlin/com/revenuecat/purchases/uifixtures/internal/AssetUrlCollectorTest.kt
@@ -1,4 +1,4 @@
-package com.revenuecat.purchases.paywallfixtures.internal
+package com.revenuecat.purchases.uifixtures.internal
 
 import org.json.JSONObject
 import kotlin.test.Test
